metadata-ingestion/requirements.txt
agate==1.9.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
babel==2.17.0
backports.tarfile==1.2.0
boto3==1.37.15
botocore==1.37.15
cached-property==1.5.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
collate-data-diff==0.11.3
collate-sqllineage==1.6.8
colorama==0.4.6
cryptography==44.0.2
daff==1.3.46
dbt-adapters==1.7.0
dbt-common==1.10.0
dbt-core==1.8.8
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
diff_cover==9.2.0
dnspython==2.7.0
dsnparse==0.1.15
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-secret-manager==2.19.0
google-crc32c==1.7.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.10
mdurl==0.1.2
memory-profiler==0.61.0
minimal-snowplow-tracker==0.0.2
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
networkx==3.2.1
numpy==2.0.2
openmetadata-ingestion==1.6.6.1
ordered-set==4.1.0
packaging==24.2
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
SecretStorage==3.3.3
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
sqlfluff==3.3.1
sqlparse==0.5.0
starlette==0.46.1
tabulate==0.9.0
tblib==3.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
zipp==3.21.0